web-app-architecture-patterns/api-edge-and-data-patterns/graphql/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
strawberry-graphql
